Refactored TestExplorer as the current design was causing memory leaks
