+++ src/Pixel.Automation.Editor.Core/Interfaces/ITestExplorerHost.cs
@@ -1,0 +1,11 @@
+﻿using Caliburn.Micro;
+
+namespace Pixel.Automation.Editor.Core.Interfaces
+{
+    /// <summary>
+    /// Host toolbar panel for the <see cref="ITestExplorer"/> views
+    /// </summary>
+    public interface ITestExplorerHost : IToolBox, IConductor
+    {
+    }
+}
