+++ src/Pixel.Automation.TestExplorer.ViewModels/EditTestCaseViewModel.cs
@@ -9,14 +9,24 @@ using System;
 
 namespace Pixel.Automation.TestExplorer.ViewModels
 {
+    /// <summary>
+    /// View Model used for editing the details of a <see cref="TestCase"/>
+    /// </summary>
     public class EditTestCaseViewModel : SmartScreen
     {
         private readonly ILogger logger = Log.ForContext<EditTestCaseViewModel>();
         private readonly TestCaseViewModel testCase;
         private readonly IEnumerable<string> existingTestCases;
 
+        /// <summary>
+        /// A clone of TestCaseViewModel where edits are made.
+        /// Actual copy will be replaced with edited copy on user confirmation after edit is done.
+        /// </summary>
         public TestCaseViewModel CopyOfTestCase { get; }
 
+        /// <summary>
+        /// DisplayName for the TestCase
+        /// </summary>
         public string TestCaseDisplayName
         {
             get => CopyOfTestCase.DisplayName;
@@ -28,6 +38,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Description for the TestCase
+        /// </summary>
         public string TestCaseDescrpition
         {
             get => CopyOfTestCase.Description;
@@ -38,44 +51,73 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Indicates if TestCase is muted.
+        /// A muted TestCase is not executed.
+        /// </summary>
         public bool IsMuted
         {
             get => CopyOfTestCase.IsMuted;
             set => CopyOfTestCase.IsMuted = value;
         }
 
+        /// <summary>
+        /// Controls the execution speed of the TestCase
+        /// </summary>
         public int DelayFactor
         {
             get => (100 - CopyOfTestCase.DelayFactor);
             set => CopyOfTestCase.DelayFactor = (100 - value);
         }
 
-
+        /// <summary>
+        /// Order of the TestCase withing a TestFixture.
+        /// Order determines the order of execution of TestCase
+        /// </summary>
         public int Order
         {
             get => CopyOfTestCase.Order;
             set => CopyOfTestCase.Order = value;
         }
 
+        /// <summary>
+        /// <see cref="Priority"/> for a TestCase.
+        /// This can be used to query the TestCases to be executed and doesn't have any significance
+        /// during the actual execution of TestCase
+        /// </summary>
         public Priority Priority
         {
             get => CopyOfTestCase.Priority;
             set => CopyOfTestCase.Priority = value;
         }
 
+        /// <summary>
+        /// A collection of (key,value) tags associated with TestCase. This can be used to filter the TestCases to be executed
+        /// and doesn't have any significance during the actual execution of TestCase.
+        /// </summary>
         public TagCollectionViewModel TagCollection { get; private set; } = new TagCollectionViewModel();
 
+        /// <summary>
+        /// constructor
+        /// </summary>
+        /// <param name="testCaseVM"></param>
+        /// <param name="existingTestCases"></param>
         public EditTestCaseViewModel(TestCaseViewModel testCaseVM, IEnumerable<TestCaseViewModel> existingTestCases)
         {
             this.testCase = testCaseVM;
             this.existingTestCases = existingTestCases.Select(s => s.DisplayName);
-            this.CopyOfTestCase = new TestCaseViewModel(testCaseVM.TestCase.Clone() as TestCase, null);
+            this.CopyOfTestCase = new TestCaseViewModel(testCaseVM.TestCase.Clone() as TestCase);
             foreach (var tag in testCaseVM.Tags)
             {
                 this.TagCollection.Add(new TagViewModel(tag));
             }
         }
 
+        /// <summary>
+        /// Apply the edits done of temporary copy to the actual instance of TestCase
+        /// and close the screen.
+        /// </summary>
+        /// <returns></returns>
         public async Task Save()
         {
             try
@@ -105,6 +147,10 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Close the screen without applying any edits done.
+        /// </summary>
+        /// <returns></returns>
         public async Task Cancel()
         {
             await this.TryCloseAsync(false);
@@ -113,14 +159,16 @@ namespace Pixel.Automation.TestExplorer.ViewModels
 
         #region Validation
 
+        /// <inheritdoc/>
         public override bool ShowModelErrors => HasErrors && propertyErrors.ContainsKey(nameof(TagCollection)) && propertyErrors[nameof(TagCollection)].Count() > 0;
 
+        /// <inheritdoc/>
         public override void DismissModelErrors()
         {
             ClearErrors(nameof(TagCollection));
             NotifyOfPropertyChange(() => ShowModelErrors);
         }
-
+       
         private bool Validate()
         {
             ValidateProperty(nameof(TestCaseDisplayName));
