+++ src/Pixel.Automation.TestExplorer.ViewModels/MockTestExplorerViewModel.cs
@@ -1,0 +1,11 @@
+﻿using Caliburn.Micro;
+
+namespace Pixel.Automation.TestExplorer.ViewModels
+{
+    /// <summary>
+    /// Dummy screen to show in TestExplorer host when there is no automation project open and active
+    /// </summary>
+    public class MockTestExplorerViewModel : Screen
+    {
+    }
+}
