+++ src/Pixel.Automation.TestExplorer.ViewModels/TestCaseViewModel.cs
@@ -1,36 +1,50 @@
-﻿using Caliburn.Micro;
-using Pixel.Automation.Core;
+﻿using Pixel.Automation.Core;
 using Pixel.Automation.Core.Enums;
 using Pixel.Automation.Core.Models;
 using Pixel.Automation.Core.TestData;
-using Pixel.Automation.Editor.Core;
 using System;
 using System.Collections.ObjectModel;
 
 namespace Pixel.Automation.TestExplorer.ViewModels
 {
+    /// <summary>
+    /// View Model for <see cref="TestCase"/>
+    /// </summary>
     public class TestCaseViewModel : NotifyPropertyChanged
-    {
-        private IEventAggregator eventAggregator;
-
+    {    
+        /// <summary>
+        /// Underlying Model for the View
+        /// </summary>
         public TestCase TestCase { get; }    
 
-        public TestCaseViewModel(TestCase testCase, IEventAggregator eventAggregator)
+        /// <summary>
+        /// constructor
+        /// </summary>
+        /// <param name="testCase"></param>
+        public TestCaseViewModel(TestCase testCase)
         {
-            this.TestCase = testCase;
-            this.eventAggregator = eventAggregator;
+            this.TestCase = testCase;           
         }
 
+        /// <summary>
+        /// Identifier of the TestCase
+        /// </summary>
         public string Id
         {
             get => TestCase.Id;           
         }
 
+        /// <summary>
+        /// Identifier of the TestFixture to which TestCase belongs
+        /// </summary>
         public string FixtureId
         {
             get => TestCase.FixtureId;            
         }
 
+        /// <summary>
+        /// DisplayName for the TestCase
+        /// </summary>
         public string DisplayName
         {
             get => TestCase.DisplayName;
@@ -41,18 +55,28 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Description for the TestCase
+        /// </summary>
         public string Description
         {
             get => TestCase.Description;
             set => TestCase.Description = value;
         }
 
+        /// <summary>
+        /// A collection of (key,value) tags associated with TestCase. This can be used to filter the TestCases to be executed
+        /// and doesn't have any significance during the actual execution of TestCase.
+        /// </summary>
         public TagCollection Tags
         {
             get => TestCase.Tags;           
         }
 
-
+        /// <summary>
+        /// Order of the TestCase withing a TestFixture.
+        /// Order determines the order of execution of TestCase
+        /// </summary>
         public int Order
         {
             get => TestCase.Order;
@@ -63,6 +87,10 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Indicates if TestCase is muted.
+        /// A muted TestCase is not executed.
+        /// </summary>
         public bool IsMuted
         {
             get => TestCase.IsMuted;
@@ -73,12 +101,20 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Controls the execution speed of the TestCase
+        /// </summary>
         public int DelayFactor
         {
             get => TestCase.DelayFactor;
             set => TestCase.DelayFactor = value;
         }
 
+        /// <summary>
+        /// <see cref="Priority"/> for a TestCase.
+        /// This can be used to query the TestCases to be executed and doesn't have any significance
+        /// during the actual execution of TestCase
+        /// </summary>
         public Priority Priority
         {
             get => TestCase.Priority;
@@ -89,6 +125,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// TestCaseEntity
+        /// </summary>
         public Entity TestCaseEntity
 
         {
@@ -96,13 +135,18 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             set => TestCase.TestCaseEntity = value;
         }
 
+        /// <summary>
+        /// Initialization ScriptFile for the TestCase
+        /// </summary>
         public string ScriptFile
         {
             get => TestCase.ScriptFile;
             set => TestCase.ScriptFile = value;
         }
 
-
+        /// <summary>
+        /// Identifier of the <see cref="TestDataSource"/> of the TestCase
+        /// </summary>
         public string TestDataId
         {
             get => TestCase.TestDataId;
@@ -114,6 +158,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// MetaData identifying all the <see cref="PrefabProject"/> used within this TestCase
+        /// </summary>
         public PrefabReferences PrefabReferences
         {
             get => TestCase.PrefabReferences;
@@ -121,6 +168,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
         }
 
         bool isSelected;
+        /// <summary>
+        /// Indicates if the TestCase is currently selected in TestExplorer
+        /// </summary>
         public bool IsSelected
         {
             get => isSelected;
@@ -132,6 +182,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
         }
 
         bool isOpenForEdit;
+        /// <summary>
+        /// Indicates if the TestCase is open for edit
+        /// </summary>
         public bool IsOpenForEdit
         {
             get => isOpenForEdit;
@@ -143,12 +196,18 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Guard method for determining whether the TestCase can be opened for edit
+        /// </summary>
         public bool CanOpenForEdit
         {
             get => !isOpenForEdit && !string.IsNullOrEmpty(TestDataId);
         }
 
         bool isRunning;
+        /// <summary>
+        /// Indicates if the TestCase is currently being executed
+        /// </summary>
         public bool IsRunning
         {
             get => isRunning;
@@ -160,6 +219,10 @@ namespace Pixel.Automation.TestExplorer.ViewModels
         }
 
         bool isVisible = true;
+        /// <summary>
+        /// Indicates whether the TestCase is visible in TestExplorer.
+        /// This is controlled by the filter applied on TestExplorer
+        /// </summary>
         public bool IsVisible
         {
             get => isVisible;
@@ -173,19 +236,30 @@ namespace Pixel.Automation.TestExplorer.ViewModels
 
         [NonSerialized]
         ObservableCollection<TestResult> testResults = new ObservableCollection<TestResult>();
+        /// <summary>
+        /// Stores the TestResult from last execution. There can be more then one result based on the
+        /// number of records in <see cref="TestDataSource"/>
+        /// </summary>
         public ObservableCollection<TestResult> TestResults
         {
             get => testResults;
             set => testResults = value;
         }
 
-
+        /// <summary>
+        /// Set the TestDataSourdce identifier for this TestCase.
+        /// TestDataSource is associated by dragging a TestDataSource to a TestCase
+        /// </summary>
+        /// <param name="testDataSourceId"></param>
         public void SetTestDataSource(string testDataSourceId)
         {
-            this.TestDataId = testDataSourceId;
-            this.eventAggregator.PublishOnBackgroundThreadAsync(new TestCaseUpdatedEventArgs(this.TestCase));
+            this.TestDataId = testDataSourceId;           
         }
 
+        /// <summary>
+        /// Determine whether this TestCase should be visible on the TestExplorer
+        /// </summary>
+        /// <param name="filterText"></param>
         public void UpdateVisibility(string filterText)
         {
             if(string.IsNullOrEmpty(filterText))
