+++ src/Pixel.Automation.TestExplorer.ViewModels/TestFixtureViewModel.cs
@@ -5,22 +5,41 @@ using System.Linq;
 
 namespace Pixel.Automation.TestExplorer.ViewModels
 {
+    /// <summary>
+    /// View Model for <see cref="TestFixture"/>
+    /// </summary>
     public class TestFixtureViewModel : NotifyPropertyChanged
     {
+        /// <summary>
+        /// Underlying Model for the view
+        /// </summary>
         public TestFixture TestFixture { get; }
 
+        /// <summary>
+        /// A colllection of TestCases beloning to this TestFixture
+        /// </summary>
         public ObservableCollection<TestCaseViewModel> Tests { get; set; } = new ObservableCollection<TestCaseViewModel>();
 
+        /// <summary>
+        /// constructor
+        /// </summary>
+        /// <param name="testFixture"></param>
         public TestFixtureViewModel(TestFixture testFixture)
         {
             this.TestFixture = testFixture;           
         }
 
+        /// <summary>
+        /// Identifier for the Fixture
+        /// </summary>
         public string Id
         {
             get => TestFixture.Id;           
         }       
 
+        /// <summary>
+        /// DisplayName for the Fixture
+        /// </summary>
         public string DisplayName
         {
             get => TestFixture.DisplayName;
@@ -31,12 +50,19 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Description for the Fixture
+        /// </summary>
         public string Description
         {
             get => TestFixture.Description;
             set => TestFixture.Description = value;
         }
 
+        /// <summary>
+        /// Indicates if the Fixture is muted.
+        /// None of the TestCases belonging to a muted fixture will be executed.
+        /// </summary>
         public bool IsMuted
         {
             get => TestFixture.IsMuted;
@@ -47,7 +73,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
-
+        /// <summary>
+        /// Order of execution of the TestFixture amongst a group of other Fixtures.
+        /// </summary>
         public int Order
         {
             get => TestFixture.Order;
@@ -58,29 +86,46 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// A collection of (key,value) tags associated with TestFixture. This can be used to filter the TestFixtures to be executed
+        /// and doesn't have any significance during the actual execution of TestCases belonging to a TestFixture.
+        /// </summary>
         public TagCollection Tags
         {
             get => TestFixture.Tags;         
         }
 
+        /// <summary>
+        /// Category to which TestFixture belongs. This can be any user defined value and is used to filter the TestFixtures to be executed
+        /// and doesn't have any significance during the actual execution of the TestCases belonging to a TestFixture.
+        /// </summary>
         public string Category
         {
             get => TestFixture.Category;
             set => TestFixture.Category = value;
         }
 
+        /// <summary>
+        /// Controls the execution speed of the TestCases belonging to a TestFixture
+        /// </summary>
         public int DelayFactor
         {
             get => TestFixture.DelayFactor;
             set => TestFixture.DelayFactor = value;
         }
         
+        /// <summary>
+        /// TestFixtureEntity
+        /// </summary>
         public Entity TestFixtureEntity
         {
             get => TestFixture.TestFixtureEntity;
             set => TestFixture.TestFixtureEntity = value;
         }
 
+        /// <summary>
+        /// Initialization Script file for a TestFixture
+        /// </summary>
         public string ScriptFile
         {
             get => TestFixture.ScriptFile;
@@ -89,6 +134,9 @@ namespace Pixel.Automation.TestExplorer.ViewModels
 
 
         bool isOpenForEdit;
+        /// <summary>
+        /// Indicates if the TestFixture is open for edit
+        /// </summary>
         public bool IsOpenForEdit
         {
             get => isOpenForEdit;
@@ -100,12 +148,18 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Guard method to determine whether the TestFixture can be opened for edit
+        /// </summary>
         public bool CanOpenForEdit
         {
             get => !isOpenForEdit;
         }
 
         bool isSelected;
+        /// <summary>
+        /// Indicates if the TestFixture is currently selected in TestExplorer
+        /// </summary>
         public bool IsSelected
         {
             get => isSelected;
@@ -117,6 +171,10 @@ namespace Pixel.Automation.TestExplorer.ViewModels
         }
 
         bool isVisible = true;
+        /// <summary>
+        /// Indicates whether the TestFixture is visible in TestExplorer.
+        /// This is controlled by the filter applied on TestExplorer
+        /// </summary>
         public bool IsVisible
         {
             get => isVisible || this.Tests.Any(t => t.IsVisible);
@@ -127,6 +185,10 @@ namespace Pixel.Automation.TestExplorer.ViewModels
             }
         }
 
+        /// <summary>
+        /// Determine whether this TestFixture should be visible on the TestExplorer
+        /// </summary>
+        /// <param name="filterText"></param>
         public void UpdateVisibility(string filterText)
         {
             foreach (var test in this.Tests)
