+++ src/Pixel.Automation.TestExplorer.Views/TestExplorerHostView.xaml
@@ -1,0 +1,13 @@
+﻿<UserControl x:Class="Pixel.Automation.TestExplorer.Views.TestExplorerHostView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Pixel.Automation.TestExplorer.Views"
+             xmlns:cal="http://www.caliburnproject.org"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <ContentControl cal:View.Model="{Binding ActiveItem}"/>
+    </Grid>
+</UserControl>
